Make the TemplatedDictionary objects picklable
The jinja2 SandboxedEnvironment objects are not guaranteed to be
pickle-serializable (even though they practically are).  So let's rather
store the environment elsewhere than in the TemplatedDictionary object
itself to keep it serializable.

Nb. TemplatedDictionary is just a wrapper around dict, which is quite
commonly subject of (de)serialization (e.g. Mock, the main user of
templated-dictionary library does that through the
concurrent.futures.ProcessPoolExecutor logic).

It would be disadvantageous to introduce additional internal arguments
into the object itself (considering we already have the __jinja_expand
internal).

Having a per-application shared environment is kind of expected anyway
(and even more optimal):

    https://jinja.palletsprojects.com/en/3.0.x/api/#basics

Note that this fix is mostly needed for the RHEL 8 environment with
jinja2 v2.10.1; newer versions made the environments (namely the
included filters) picklable: https://github.com/pallets/jinja/pull/1613